Configured Babel, refactored promise chains to async/await
